finished services changes
